requirements-native.txt
pyinstaller>=6.19.0
pillow>=11.3.0
